test: refactor using enzyme shallow
